DABOM-251 refactor: policy Page 수정
